Add support for python3.9
## Why?
Python3.9 will be EOL'ed in October 2025, which means
an additional 18 months of support.

https://endoflife.date/python

I'd like to start using autotransform in a repo that
is currently using python3.9 and I'd like to avoid
a separate virtual environment, requirements file,
etc.

## What?
- Modify setup.py to allow python3.9
- Bump version to 1.1.1a14
- Update references in `validator/base.py` to
handle the `@dataclass` decorator not supporting
the `kw_only` arg in python3.9.
- Update references to union types using the `|`
syntax to use `Union[]` instead.